Refactor service module
